+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-29T10:37:53.086Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-29T11:58:31.816Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")