+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1903,7 +1903,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KSSN03TVD1DW4E2536161PF3"
+    id: Optional[str] = "01KSSSKVA2NH0CDP1T4GKYSZKS"
 
     description: Optional[str] = None